pyproject.toml
[project]
name = "little-turtle"
version = "0.1.0"
description = ""
authors = [{ email = "pavel.karpovich@outlook.com", name = "Pavel Karpovich" }]
readme = "README.md"
requires-python = ">= 3.12"
dependencies = [
    "langchain>=0.3.26",
    "python-dotenv>=1.0.1",
    "openai>=1.23.6",
    "telethon>=1.35.0",
    "dependency-injector-fork>=4.42.1",
    "aiogram>=3.5.0",
    "structlog>=24.1.0",
    "sentry-sdk>=2.0.1",
    "apscheduler>=3.10.4",
    "redis>=5.0.3",
    "jinja2>=3.1.3",
    "langfuse>=2.27.2",
    "langchain-openai>=0.3.27",
    "pytz>=2025.2",
    "pydantic>=2.11.7",
    "prompt-template>=1.2.0",
    "anthropic>=0.57.1",
    "arize-phoenix-otel>=0.12.1",
    "openinference-instrumentation-openai>=0.1.30",
    "openinference-instrumentation-anthropic>=0.1.18",
    "arize-phoenix-client>=1.13.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.2"
]
